requirements.txt
aiohttp==3.8.4
python-telegram-bot==13.7
Flask==2.3.2
python-dotenv==1.0.1
